chore: type 분리
