--- apps/mobile/type/selectedDateType.ts
@@ -1,13 +1,0 @@
-export interface SelectedDateType {
-  startYear: number;
-  startMonth: number;
-  startDate: number;
-  endYear: number;
-  endMonth: number;
-  endDate: number;
-}
-
-export interface SelectedTimeType {
-  startTime: string;
-  endTime: string;
-}
